Android Application to perform rodeo event where accelerometer events are registered
Android Application to perform rodeo event where accelerometer events
are registered. The game is governed by certain set of rules in horse
circling the barrel, penalty,etc. It also aims at providing a better
user interface by providing provisions for pausing the game, viewing
highscore & game play instructions. The project followed Agile
development where each iteration time-boxed by a week. After the first
two iteration 90% of game requirements in terms of User interface,
gaming rules and constraints, provisions for Users have been discussed
and 20% of code has been programmed and in subsequent iterations have
sucessfully completed the project.

Rules/Constraints to build the gaming app:
A barrel race is a rodeo event in which the rider starts at a gate and
must ride completely around three (sometimes four) barrels.  That is, he
or she must circle each of the barrels.  The object is to get the
fastest time without knocking over any of the barrels.  In the game you
will write, your “horse” starts out at the gate, and you must maneuver
him around the three barrels, as shown in the diagram, and back through
the gate. You do this by tilting your device one way or another.  The
course should be square and at the bottom of your screen, leaving room
for a start button (not shown) and the time.  Touching any of the
barrels means you lose, and must start over.  Hitting the fences around
the course will add 5 seconds to your time, but you get to continue the
race. The circles in the center arranged in a triangle are the barrels.
The black circle at the bottom is your horse at the starting gate, and
the time is shown as minutes, seconds, and hundredths.  The timer starts
when you press the start button and ends when your horse is entirely
through the gate.

Software tools: Eclipse IDE.